pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kii-readiness-mapper"
version = "0.1.0"
description = "Tool for self-assessment of readiness for Russian Critical Information Infrastructure (CII) status"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
  {name = "Run As Daemon", email = "info@run-as-daemon.ru"}
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
  "typer>=0.9.0",
  "fastapi>=0.100.0",
  "uvicorn>=0.22.0",
  "jinja2>=3.1.0",
  "pyyaml>=6.0",
  "pydantic>=2.0.0",
  "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.0.270",
  "mypy>=1.0.0",
  "bandit>=1.7.0",
  "types-PyYAML",
  "httpx",
]

[project.scripts]
kii-readiness = "kii_readiness_mapper.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
